binize int enforce in ccg()

+++ generalephys.py
@@ -1339,7 +1339,7 @@ def ccg(train1,train2,binrange,binsize):
                     diffs.extend([float(spiketime_train1) - float(spiketime_train2)])
                     count+=1
         diffs = np.array(diffs)*-1
-        hist,edges = np.histogram(diffs,bins=(binrange[1]-binrange[0])/binsize,range=binrange)
+        hist,edges = np.histogram(diffs,bins=int((binrange[1]-binrange[0])/binsize),range=binrange)
         return (hist / float(len(train1)))*100,edges
         #return (hist / float(len(train1)*len(train2) / 2.)  ,edges)
         #return ((hist / (len(train1) * len(train2)) / 2.)*100 * binsize,edges)
